RAC-379: Integrate summary in execution detail
